+++ .flowbite-react/class-list.json
@@ -1,33 +1,32 @@
 [
   "-4px",
+  "2xl",
+  "3xl",
+  "4xl",
+  "5xl",
+  "6xl",
+  "7xl",
   "absolute",
+  "bg-blue-700",
+  "bg-cyan-700",
+  "bg-gray-700",
+  "bg-gray-800",
   "bg-gray-900",
+  "bg-gray-900/50",
+  "bg-green-700",
+  "bg-indigo-700",
+  "bg-lime-700",
+  "bg-pink-700",
+  "bg-primary-700",
+  "bg-purple-700",
+  "bg-red-700",
+  "bg-teal-700",
+  "bg-transparent",
   "bg-white",
+  "bg-yellow-400",
   "border",
-  "border-gray-200",
-  "dark:bg-gray-700",
-  "dark:border-none",
-  "dark:text-white",
-  "font-medium",
-  "h-2",
-  "inline-block",
-  "invisible",
-  "opacity-0",
-  "px-3",
-  "py-2",
-  "relative",
-  "rotate-45",
-  "rounded-lg",
-  "shadow-xs",
-  "text-gray-900",
-  "text-sm",
-  "text-white",
-  "transition-opacity",
-  "w-2",
-  "w-fit",
-  "z-10",
-  "z-20"
-],
+  "border-b",
+  "border-b-0",
   "border-blue-700",
   "border-cyan-700",
   "border-gray-200",
